requirements.txt
# requirements.txt
# Core
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.0

# Redis
redis==5.0.1

# Telegram Bot
aiogram==3.3.0

# Telegram Client (for transfer)
telethon==1.34.0
cryptography==41.0.7

# Web Framework
fastapi==0.108.0
uvicorn[standard]==0.25.0
python-multipart==0.0.6

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# WebSocket
websockets==12.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Logging
loguru==0.7.2
